encrypted-websocket-1vdnak/pyproject.toml
[project]
name = "encrypted-websocket-1vdnak"
version = "0.1.0"
description = ""
authors = [
    {name = "arunanksharan",email = "arunank.sharan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "livekit-agents (>=1.0.0,<1.1.0)",
    "livekit-plugins-openai (>=1.0.11,<2.0.0)",
    "livekit-plugins-deepgram (>=1.0.11,<2.0.0)",
    "livekit-plugins-silero (>=1.0.11,<2.0.0)",
    "livekit-plugins-turn-detector (>=1.0.11,<2.0.0)",
    "livekit-plugins-noise-cancellation (>=0.2.1,<0.3.0)",
    "livekit-plugins-elevenlabs (>=1.0.11,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
